Mobilanpassning

+++ participant/view_alchemy_supplier.php
@@ -20,7 +20,7 @@ if ($_SERVER["REQUEST_METHOD"] == "GET") {
 $role = Role::loadById($RoleId);
 $person = $role->getPerson();
 
-if ($person->UserId != $current_user->Id) {
+if ($person->Id != $current_person->Id) {
     header('Location: index.php'); //Inte din karaktär
     exit;
 }
@@ -36,65 +36,80 @@ $supplier = Alchemy_Supplier::getForRole($role);
 
 include 'navigation.php';
 ?>
+<style>
+input {
+  width: 60px;
+  text-align: right; 
+}
+</style>
 <script src="../javascript/setringredientamount_ajax.js"></script>
 
-	<div class="content">
-		<h1><?php echo "Löjverist $role->Name";?></h1>
+	<div class='itemselector'>
+		<div class="header">
+
+			<i class="fa-solid fa-leaf"></i>
+			<?php echo "Löjverist $role->Name";?>
+		</div>
 		
+   		<div class='itemcontainer'>
+           	<div class='itemname'>Workshop datum</div>
+			<?php 
+			if ($supplier->hasDoneWorkshop()) echo $supplier->Workshop; 
+			    else echo showParticipantStatusIcon(false, "Du har inte deltagit i workshop om alkemi/lövjeri");
+		    ?>			
+		</div>
+	</div>
+	<div class='itemselector'>
+		<div class="header">
 
-		<div>
-    		<table>
-    			<tr>
-    				<td>Workshop datum</td>
-     				<td><?php 
-     				    echo showStatusIcon($supplier->hasDoneWorkshop())." ";
-     				
-    				    if ($supplier->hasDoneWorkshop()) echo $supplier->Workshop; 
-    				    else echo "Du har inte deltagit i workshop om alkemi/lövjeri.";
-    				    ?></td>
-    			</tr>
-    		</table>
+			<i class="fa-solid fa-leaf"></i> Ingredienser till <?php echo $current_larp->Name?>
+		</div>
+   		<div class='itemcontainer'>
+		Lägg till de ingredienser som du tänker ta med till lajvet. Arrangörerna godkänner sedan att du får ta med ingrediensen och hur många du får ta med.<br>
+		<a href='alchemy_all_ingredients.php?RoleId=<?php echo $role->Id?>'>Visa alla ingredienser som finns / Välj ingredienser att ta med</a><br>
+		</div>
+   		<div class='itemcontainer'>
 
-			<h2>Ingredienser till <?php echo $current_larp->Name?></h2>
-			<p>Lägg till de ingredienser som du tänker ta med till lajvet. Arrangörerna godkänner sedan att du får ta med ingrediensen och hur många du får ta med.</p>
-				<a href='alchemy_all_ingredients.php?RoleId=<?php echo $role->Id?>'>Visa alla ingredienser som finns / Välj ingredienser att ta med</a><br><br>
+			<div  style='display:table'>
 
+ 
 			<?php 
 			$amounts = $supplier->getIngredientAmounts($current_larp);
 			if (empty($amounts)) {
 			    echo "Inga valda ingredenser, än.";
 			} else {
 			    echo "Så länge arrangörerna inte har godkänt att du får ta med dig ingrediensen och hur många du tar med dig, så kan du ändra hur många du vill ha med dig och du kan även ta bort den helt från din lista. Om du vill ändra efter att den är godkänd behöver du kontakta arrangörerna.";
-				echo "<table class='data'>";
-				echo "<tr><th>Ingrediens</th><th>Antal</th><th>Nivå</th><th>Ingrediens/<br>Katalysator</th><th>Essenser</th><th>Off-ingrediens</th><th>Godkänd/<br>Ännu inte godkänd</th><th></th></tr>";
+
+			    echo "<table class='participant_table' style='width:100%;'>";
+				echo "<tr><th>Ingrediens</th><th>Essenser</th><th>Antal</th><th>Godkänd/<br>Ännu inte godkänd</th><th></th></tr>";
 				foreach ($amounts as $amount) {
 				    $ingredient = $amount->getIngredient();
 				    echo "<tr>";
-				    echo "<td>$ingredient->Name</td>\n";
-				    echo "<td>";
-				    if ($amount->isApproved()) echo $amount->Amount;
-				    else {
-				        echo "<input type='number' id='$amount->Id' min='1' value='$amount->Amount' onchange='saveAmount(this)'>";
-				    }
-				    echo "</td>";
-				    echo "<td>$ingredient->Level</td>\n";
-				    echo "<td>";
+				    echo "<td>$ingredient->Name<br>";
 				    if ($ingredient->isCatalyst()) echo "Katalysator";
 				    else echo "Ingrediens";
+				    echo "<br>";
+				    echo "Nivå $ingredient->Level<br>\n";
 				    echo "</td>\n";
 				    echo "<td>";
 				    if ($ingredient->isIngredient()) {
 				        echo $ingredient->getEssenceNames();
 				    }
 				    echo "</td>\n";
-				    
-				    echo "<td>$ingredient->ActualIngredient</td>\n";
 				    echo "<td>";
-				    echo showStatusIcon($amount->isApproved());
+				    if ($amount->isApproved()) echo $amount->Amount;
+				    else {
+				        echo "<input type='number' id='$amount->Id' min='1' value='$amount->Amount' onchange='saveAmount(this)' maxlength='3' size='4'>";
+				    }
+				    echo "</td>";
+				    
+				    echo "<td>";
+				    if ($amount->isApproved()) echo showStatusIcon(true);
+				    else echo showQuestionmarkIcon();
 				    
 				    echo "</td>\n";
 				    echo "<td>";
-				    echo "<a href='view_alchemy_supplier.php?operation=delete&supplierIngredientId=$amount->Id&id=$role->Id'><i class='fa-solid fa-trash'></i>";
+				    if (!$amount->isApproved()) echo "<a href='view_alchemy_supplier.php?operation=delete&supplierIngredientId=$amount->Id&id=$role->Id'><i class='fa-solid fa-trash'></i>";
 				    
 				    
 				    echo "</td>\n";
@@ -104,8 +119,9 @@ include 'navigation.php';
 			}
 			?>
 
-
 		</div>
+		</div>
+	</div>
